Add files via upload
This Python data analysis project involves setting up a cloud-based environment on AWS. To manage access, an IAM user will be created with appropriate permissions for EC2 and S3. This user will be used to connect to a remote Gitbash instance via SSH for code development. While direct SSH connections without an IAM user are possible, a hybrid approach is preferred for better security and management. The project will leverage an EC2 instance for computational tasks and potentially utilize S3 for data storage based on project requirements.